requirements.txt
python-social-auth
social-auth-app-django
configparser
social-auth-core[openidconnect]
braintree
channels
channels_redis
WeasyPrint==0.26
celery
django-ckeditor
django-import-export
python-dateutil
mysqlclient==1.3.5
djangorestframework
django-debug-toolbar
openpyxl==2.4.9
redis
async-timeout==2.0.1
gunicorn3
celery
django-celery
